requirements.txt
django==5.0.1
djangorestframework
djangorestframework-simplejwt
pyyaml
requests
django-cors-headers
psycopg2-binary
bcrypt==4.1.2